Move the state override types to their appropriate modules
